+++ sshd-common/src/main/java/org/apache/sshd/common/cipher/BuiltinCiphers.java
@@ -46,32 +46,32 @@ import org.apache.sshd.common.util.ValidateUtils;
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 public enum BuiltinCiphers implements CipherFactory {
-    none(Constants.NONE, 0, 0, "None", "None") {
+    none(Constants.NONE, 0, 0, "None", 0, "None") {
         @Override
         public Cipher create() {
             return new CipherNone();
         }
     },
-    aes128cbc(Constants.AES128_CBC, 16, 16, "AES", "AES/CBC/NoPadding"),
-    aes128ctr(Constants.AES128_CTR, 16, 16, "AES", "AES/CTR/NoPadding"),
-    aes192cbc(Constants.AES192_CBC, 16, 24, "AES", "AES/CBC/NoPadding"),
-    aes192ctr(Constants.AES192_CTR, 16, 24, "AES", "AES/CTR/NoPadding"),
-    aes256cbc(Constants.AES256_CBC, 16, 32, "AES", "AES/CBC/NoPadding"),
-    aes256ctr(Constants.AES256_CTR, 16, 32, "AES", "AES/CTR/NoPadding"),
-    arcfour128(Constants.ARCFOUR128, 8, 16, "ARCFOUR", "RC4") {
+    aes128cbc(Constants.AES128_CBC, 16, 16, "AES", 128, "AES/CBC/NoPadding"),
+    aes128ctr(Constants.AES128_CTR, 16, 16, "AES", 128, "AES/CTR/NoPadding"),
+    aes192cbc(Constants.AES192_CBC, 16, 24, "AES", 192, "AES/CBC/NoPadding"),
+    aes192ctr(Constants.AES192_CTR, 16, 24, "AES", 192, "AES/CTR/NoPadding"),
+    aes256cbc(Constants.AES256_CBC, 16, 32, "AES", 256, "AES/CBC/NoPadding"),
+    aes256ctr(Constants.AES256_CTR, 16, 32, "AES", 256, "AES/CTR/NoPadding"),
+    arcfour128(Constants.ARCFOUR128, 8, 16, "ARCFOUR", 128, "RC4") {
         @Override
         public Cipher create() {
-            return new BaseRC4Cipher(getIVSize(), getBlockSize());
+            return new BaseRC4Cipher(getIVSize(), getKdfSize(), getKeySize());
         }
     },
-    arcfour256(Constants.ARCFOUR256, 8, 32, "ARCFOUR", "RC4") {
+    arcfour256(Constants.ARCFOUR256, 8, 32, "ARCFOUR", 256, "RC4") {
         @Override
         public Cipher create() {
-            return new BaseRC4Cipher(getIVSize(), getBlockSize(), getKeySize());
+            return new BaseRC4Cipher(getIVSize(), getKdfSize(), getKeySize());
         }
     },
-    blowfishcbc(Constants.BLOWFISH_CBC, 8, 16, "Blowfish", "Blowfish/CBC/NoPadding"),
-    tripledescbc(Constants.TRIPLE_DES_CBC, 8, 24, "DESede", "DESede/CBC/NoPadding");
+    blowfishcbc(Constants.BLOWFISH_CBC, 8, 16, "Blowfish", 128, "Blowfish/CBC/NoPadding"),
+    tripledescbc(Constants.TRIPLE_DES_CBC, 8, 24, "DESede", 192, "DESede/CBC/NoPadding");
 
     public static final Set<BuiltinCiphers> VALUES =
         Collections.unmodifiableSet(EnumSet.allOf(BuiltinCiphers.class));
@@ -81,17 +81,17 @@ public enum BuiltinCiphers implements CipherFactory {
 
     private final String factoryName;
     private final int ivsize;
-    private final int blocksize;
+    private final int kdfSize;
     private final int keysize;
     private final String algorithm;
     private final String transformation;
     private final boolean supported;
 
-    BuiltinCiphers(String factoryName, int ivsize, int blocksize, String algorithm, String transformation) {
+    BuiltinCiphers(String factoryName, int ivsize, int kdfSize, String algorithm, int keySize, String transformation) {
         this.factoryName = factoryName;
         this.ivsize = ivsize;
-        this.blocksize = blocksize;
-        this.keysize = blocksize * Byte.SIZE;
+        this.kdfSize = kdfSize;
+        this.keysize = keySize;
         this.algorithm = algorithm;
         this.transformation = transformation;
         /*
@@ -134,8 +134,8 @@ public enum BuiltinCiphers implements CipherFactory {
     }
 
     @Override
-    public int getBlockSize() {
-        return blocksize;
+    public int getKdfSize() {
+        return kdfSize;
     }
 
     @Override
@@ -150,7 +150,7 @@ public enum BuiltinCiphers implements CipherFactory {
 
     @Override
     public Cipher create() {
-        return new BaseCipher(getIVSize(), getBlockSize(), getAlgorithm(), getKeySize(), getTransformation());
+        return new BaseCipher(getIVSize(), getKdfSize(), getAlgorithm(), getKeySize(), getTransformation());
     }
 
     /**